- add JSON config to String - add config to CreateGameReq - refactor GameCreation to create game with config that given from client - fix bux that occur when player connected to webpage in first time. (null props) - edit type of config to 'long'

+++ src/main/java/com/example/springtest/RequestBody/updateConfig.java
@@ -4,15 +4,15 @@ import lombok.Data;
 
 @Data
 public class updateConfig {
-    private int rows;
-    private int cols;
-    private int initPlanMin;
-    private int initPlanSec;
-    private double initBudget;
-    private double initCenterDep;
-    private int planRevMin;
-    private int planRevSec;
-    private double revCost;
-    private double maxDep;
-    private double interestPct;
+    private long rows;
+    private long cols;
+    private long initPlanMin;
+    private long initPlanSec;
+    private long initBudget;
+    private long initCenterDep;
+    private long planRevMin;
+    private long planRevSec;
+    private long revCost;
+    private long maxDep;
+    private long interestPct;
 }
